RLSeq Results

Functions for creating and accessing the R-loop results (RL Results). These are a type of object used for holding the results of the tests implemented in RLSeq. They can be accessed using the rlresult function.

rlresult(object, resultName)

Arguments

object

RLRanges object.

resultName

Name of the result slot to access. See details.

Value

The contents of the requested slot.

Details

Slot descriptions

featureEnrichment

The tbl generated from running featureEnrich.

The structure and column descriptions are provided in detail within RLHub::feat_enrich_samples.

correlationMat

The matrix generated from running corrAnalyze.

Contains pairwise pearson correlations between all samples in RLBase and the supplied RLRanges object.

rlfsRes

The list generated from running analyzeRLFS.

See analyzeRLFS for description of structure.

noiseAnalysis

The tbl generated from running noiseAnalysis.

txFeatureOverlap

The tbl generated from running txFeatureOverlap.

geneAnnoRes

The tbl generated from running geneAnnotation.

predictRes

The list generated from running predictCondition.

rlRegionRes

The list generated from running rlRegionTest.

Examples

rlr <- readRDS(system.file("extdata", "rlrsmall.rds", package = "RLSeq"))

rlresult(rlr, "predictRes")
#> $Features
#> # A tibble: 12 × 3
#>    feature raw_value processed_value
#>    <chr>       <dbl>           <dbl>
#>  1 Z1         -0.989        3.78e+ 0
#>  2 Z2         -1.18         5.69e+ 1
#>  3 Zacf1      -0.972        4.49e- 3
#>  4 Zacf2      -1.28         5.14e+ 0
#>  5 ReW1       -0.513        4.32e+ 0
#>  6 ReW2       -1.18         7.79e+ 2
#>  7 ImW1       -2.73        -3.52e-15
#>  8 ImW2        1.04         2.09e+ 2
#>  9 ReWacf1    -1.20         1.80e+ 0
#> 10 ReWacf2    -1.25         5.80e+ 1
#> 11 ImWacf1    -0.582        2.71e-16
#> 12 ImWacf2    -1.26         4.42e+ 1
#> 
#> $Criteria
#> $Criteria$`PVal Significant`
#> [1] TRUE
#> 
#> $Criteria$`ZApex > 0`
#> [1] TRUE
#> 
#> $Criteria$`ZApex > ZEdges`
#> [1] FALSE
#> 
#> $Criteria$`Predicted 'POS'`
#> [1] FALSE
#> 
#> 
#> $prediction
#> [1] "NEG"
#>