RLSeq

Executes the RLSeq analysis workflow.

RLSeq(object, quiet = FALSE, skip = NULL, ...)

Arguments

object

An RLRanges object.

quiet

If TRUE, messages are suppressed. Default: FALSE.

skip

Analysis steps to skip. Default: NULL. See details for options.

...

Arguments passed to analyzeRLFS.

Value

An RLRanges object with results available (see rlresult).

Details

The RLSeq() function does all of the following by default:

RLFS Perm Test. Runs the analyzeRLFS function to test the enrichment of user-supplied ranges within R-loop-forming sequences. Cannot be skipped.

Predict Condition. Runs the predictCondition function to predict whether the user-supplied sample robustly maps R-loops or not. Cannot be skipped.

Noise analysis. Runs the noiseAnalyze function to analyze the signal-noise distribution within the supplied dataset. Skip with skip="noiseAnalyze".

Feature enrichment test. Runs the featureEnrich function to test the enrichment of user-supplied ranges within R-loop-relevant genomic features. Skip with skip="featureEnrich".

Transcript Feature Overlap. Runs the txFeatureOverlap function to get the overlap of transcript features and user-supplied peaks.

Correlation Analysis. Runs the corrAnalyze function to test the correlation of user-supplied R-loop signal with other samples in RLBase around "gold-standard" R-loop regions. Skip with skip="corrAnalyze".

Gene annotation. Runs the geneAnnotation function to find overlap of genes with the user-supplied ranges. Skip with skip="geneAnnotation".

R-loop Region Analysis. Runs the rlRegionTest function to find the overlap of user-supplied ranges with consensus R-loop sites (RL-Regions). Skip with skip="rlRegionTest".

Examples

# Example RLRanges
rlr <- readRDS(system.file("extdata", "rlrsmall.rds", package = "RLSeq"))

# Run RLSeq
# `useMask=FALSE`, `ntime=10`, and `skip=` for demonstration purposes here.
rlr <- RLSeq(
    rlr,
    useMask = FALSE, ntimes = 10,
    skip = c(
        "featureEnrich", "corrAnalyze", "geneAnnotation", "rlRegionTest"
    )
)
#> [1/8] RLFS Perm Test
#> [1] "Note: The minimum p-value with only 10 permutations is 0.0909090909090909. You should consider increasing the number of permutations."
#> Warning: GRanges object contains 1 out-of-bound range located on sequence
#>   chr22_KI270733v1_random. Note that ranges located on a sequence whose
#>   length is unknown (NA) or on a circular sequence are not considered
#>   out-of-bound (use seqlengths() and isCircular() to get the lengths and
#>   circularity flags of the underlying sequences). You can use trim() to
#>   trim these ranges. See ?`trim,GenomicRanges-method` for more
#>   information.
#> Warning: GRanges object contains 1 out-of-bound range located on sequence
#>   chr22_KI270733v1_random. Note that ranges located on a sequence whose
#>   length is unknown (NA) or on a circular sequence are not considered
#>   out-of-bound (use seqlengths() and isCircular() to get the lengths and
#>   circularity flags of the underlying sequences). You can use trim() to
#>   trim these ranges. See ?`trim,GenomicRanges-method` for more
#>   information.
#> Warning: GRanges object contains 1 out-of-bound range located on sequence
#>   chr22_KI270733v1_random. Note that ranges located on a sequence whose
#>   length is unknown (NA) or on a circular sequence are not considered
#>   out-of-bound (use seqlengths() and isCircular() to get the lengths and
#>   circularity flags of the underlying sequences). You can use trim() to
#>   trim these ranges. See ?`trim,GenomicRanges-method` for more
#>   information.
#> Warning: GRanges object contains 1 out-of-bound range located on sequence
#>   chr22_KI270733v1_random. Note that ranges located on a sequence whose
#>   length is unknown (NA) or on a circular sequence are not considered
#>   out-of-bound (use seqlengths() and isCircular() to get the lengths and
#>   circularity flags of the underlying sequences). You can use trim() to
#>   trim these ranges. See ?`trim,GenomicRanges-method` for more
#>   information.
#> Warning: GRanges object contains 1 out-of-bound range located on sequence
#>   chr22_KI270733v1_random. Note that ranges located on a sequence whose
#>   length is unknown (NA) or on a circular sequence are not considered
#>   out-of-bound (use seqlengths() and isCircular() to get the lengths and
#>   circularity flags of the underlying sequences). You can use trim() to
#>   trim these ranges. See ?`trim,GenomicRanges-method` for more
#>   information.
#> Warning: GRanges object contains 1 out-of-bound range located on sequence
#>   chr22_KI270733v1_random. Note that ranges located on a sequence whose
#>   length is unknown (NA) or on a circular sequence are not considered
#>   out-of-bound (use seqlengths() and isCircular() to get the lengths and
#>   circularity flags of the underlying sequences). You can use trim() to
#>   trim these ranges. See ?`trim,GenomicRanges-method` for more
#>   information.
#> Warning: GRanges object contains 1 out-of-bound range located on sequence
#>   chr22_KI270733v1_random. Note that ranges located on a sequence whose
#>   length is unknown (NA) or on a circular sequence are not considered
#>   out-of-bound (use seqlengths() and isCircular() to get the lengths and
#>   circularity flags of the underlying sequences). You can use trim() to
#>   trim these ranges. See ?`trim,GenomicRanges-method` for more
#>   information.
#> Warning: GRanges object contains 1 out-of-bound range located on sequence
#>   chr22_KI270733v1_random. Note that ranges located on a sequence whose
#>   length is unknown (NA) or on a circular sequence are not considered
#>   out-of-bound (use seqlengths() and isCircular() to get the lengths and
#>   circularity flags of the underlying sequences). You can use trim() to
#>   trim these ranges. See ?`trim,GenomicRanges-method` for more
#>   information.
#> Warning: GRanges object contains 1 out-of-bound range located on sequence
#>   chr22_KI270733v1_random. Note that ranges located on a sequence whose
#>   length is unknown (NA) or on a circular sequence are not considered
#>   out-of-bound (use seqlengths() and isCircular() to get the lengths and
#>   circularity flags of the underlying sequences). You can use trim() to
#>   trim these ranges. See ?`trim,GenomicRanges-method` for more
#>   information.
#> Warning: GRanges object contains 1 out-of-bound range located on sequence
#>   chr22_KI270733v1_random. Note that ranges located on a sequence whose
#>   length is unknown (NA) or on a circular sequence are not considered
#>   out-of-bound (use seqlengths() and isCircular() to get the lengths and
#>   circularity flags of the underlying sequences). You can use trim() to
#>   trim these ranges. See ?`trim,GenomicRanges-method` for more
#>   information.
#> Warning: GRanges object contains 1 out-of-bound range located on sequence
#>   chr22_KI270733v1_random. Note that ranges located on a sequence whose
#>   length is unknown (NA) or on a circular sequence are not considered
#>   out-of-bound (use seqlengths() and isCircular() to get the lengths and
#>   circularity flags of the underlying sequences). You can use trim() to
#>   trim these ranges. See ?`trim,GenomicRanges-method` for more
#>   information.
#> Warning: GRanges object contains 1 out-of-bound range located on sequence
#>   chr22_KI270733v1_random. Note that ranges located on a sequence whose
#>   length is unknown (NA) or on a circular sequence are not considered
#>   out-of-bound (use seqlengths() and isCircular() to get the lengths and
#>   circularity flags of the underlying sequences). You can use trim() to
#>   trim these ranges. See ?`trim,GenomicRanges-method` for more
#>   information.
#> Warning: GRanges object contains 1 out-of-bound range located on sequence
#>   chr22_KI270733v1_random. Note that ranges located on a sequence whose
#>   length is unknown (NA) or on a circular sequence are not considered
#>   out-of-bound (use seqlengths() and isCircular() to get the lengths and
#>   circularity flags of the underlying sequences). You can use trim() to
#>   trim these ranges. See ?`trim,GenomicRanges-method` for more
#>   information.
#> Warning: GRanges object contains 1 out-of-bound range located on sequence
#>   chr22_KI270733v1_random. Note that ranges located on a sequence whose
#>   length is unknown (NA) or on a circular sequence are not considered
#>   out-of-bound (use seqlengths() and isCircular() to get the lengths and
#>   circularity flags of the underlying sequences). You can use trim() to
#>   trim these ranges. See ?`trim,GenomicRanges-method` for more
#>   information.
#> Warning: GRanges object contains 1 out-of-bound range located on sequence
#>   chr22_KI270733v1_random. Note that ranges located on a sequence whose
#>   length is unknown (NA) or on a circular sequence are not considered
#>   out-of-bound (use seqlengths() and isCircular() to get the lengths and
#>   circularity flags of the underlying sequences). You can use trim() to
#>   trim these ranges. See ?`trim,GenomicRanges-method` for more
#>   information.
#> Warning: GRanges object contains 1 out-of-bound range located on sequence
#>   chr22_KI270733v1_random. Note that ranges located on a sequence whose
#>   length is unknown (NA) or on a circular sequence are not considered
#>   out-of-bound (use seqlengths() and isCircular() to get the lengths and
#>   circularity flags of the underlying sequences). You can use trim() to
#>   trim these ranges. See ?`trim,GenomicRanges-method` for more
#>   information.
#> Warning: GRanges object contains 1 out-of-bound range located on sequence
#>   chr22_KI270733v1_random. Note that ranges located on a sequence whose
#>   length is unknown (NA) or on a circular sequence are not considered
#>   out-of-bound (use seqlengths() and isCircular() to get the lengths and
#>   circularity flags of the underlying sequences). You can use trim() to
#>   trim these ranges. See ?`trim,GenomicRanges-method` for more
#>   information.
#> Warning: GRanges object contains 1 out-of-bound range located on sequence
#>   chr22_KI270733v1_random. Note that ranges located on a sequence whose
#>   length is unknown (NA) or on a circular sequence are not considered
#>   out-of-bound (use seqlengths() and isCircular() to get the lengths and
#>   circularity flags of the underlying sequences). You can use trim() to
#>   trim these ranges. See ?`trim,GenomicRanges-method` for more
#>   information.
#> [2/8] Predict Condition
#> see ?RLHub and browseVignettes('RLHub') for documentation
#> see ?RLHub and browseVignettes('RLHub') for documentation
#> [3/8] Noise Analysis
#> [4/8] Feature Enrichment Test
#> [5/8] Transcript Feature Overlap Analysis
#> see ?RLHub and browseVignettes('RLHub') for documentation
#> [6/8] Correlation Analysis
#> No coverage provided... skipping.
#> [7/8] Gene Annotation
#> [8/8] R-loop Region Analysis