Auxiliary Data

A list containing data used by RLSeq functions. It can also be useful for checking the available modes and genomes in RLSeq. See also the data-raw/auxdata.R script that was used to create it.

auxdata

Format

An object of class list of length 10.

Details

Structure

A named list containing the following entries:

db_cols

A tbl with colors associated with each database in RLHub useful for plotting. See also RLHub::annotations.

annotypes

A tbl containing the annotation databases and annotation types available from RLBase. See also RLHub::annotations.

ip_cols

A tbl containing the colors associated with each "Immunoprecipitation type" (ip_type) in RLBase. See also RLHub::rlbase_samples.

mode_cols

A tbl containing the colors associated with each R-loop mapping mode in RLHub::rlbase_samples.

heat_cols

A tbl containing the colors associated with user-supplied data and RLBase data when running corrHeatmap.

label_cols

A tbl containing the colors associated with the labels in RLBase. See also RLHub::rlbase_samples.

prediction_cols

A tbl containing the colors associated with the predictions in RLBase. See also RLHub::rlbase_samples.

available_modes

A tbl containing the modes available in RLBase and associated metadata. See also RLHub::rlbase_samples.

available_genomes

A character showing all the official UCSC genomes available for use with RLSeq. See also available_genomes.

misc_modes

A character showing the R-loop mapping modes that are lumped into the 'misc' category for simplification of plotting.

Examples

auxdata