Genome Masks

A collection of genome masks for use with analyzeRLFS. See the data-raw/genome_masks.R script for the processing steps.

genomeMasks

Format

An object of class list of length 8.

Details

Structure

genomeMasks is a named list of GRanges objects. Each entry in the list follows the naming convention: <genome>.masked, where <genome> is an official UCSC genome ID. Each entry contains a GRanges object with the masked ranges from <genome>. The genomes provided correspond to the masked genomes available in BSgenome::available.genomes.

Examples

genomeMasks
#> $apiMel2.masked
#> GRanges object with 16164 ranges and 0 metadata columns:
#>           seqnames              ranges strand
#>              <Rle>           <IRanges>  <Rle>
#>       [1]   Group1           8604-9586      *
#>       [2]   Group1         45114-45303      *
#>       [3]   Group1         74031-74436      *
#>       [4]   Group1       127756-127879      *
#>       [5]   Group1       129251-129560      *
#>       ...      ...                 ...    ...
#>   [16160]  GroupUn 320949657-320984656      *
#>   [16161]  GroupUn 320985744-321020743      *
#>   [16162]  GroupUn 321029717-321064716      *
#>   [16163]  GroupUn 321065873-321100872      *
#>   [16164]  GroupUn 321103175-321138174      *
#>   -------
#>   seqinfo: 17 sequences from apiMel2 genome
#> 
#> $canFam3.masked
#> GRanges object with 19553 ranges and 0 metadata columns:
#>           seqnames              ranges strand
#>              <Rle>           <IRanges>  <Rle>
#>       [1]     chr1               1-100      *
#>       [2]     chr1         24126-27778      *
#>       [3]     chr1         45598-45882      *
#>       [4]     chr1         89487-89496      *
#>       [5]     chr1        91795-162650      *
#>       ...      ...                 ...    ...
#>   [19549]     chrX 123311357-123312638      *
#>   [19550]     chrX 123445883-123445892      *
#>   [19551]     chrX 123776002-123776553      *
#>   [19552]     chrX 123821636-123822296      *
#>   [19553]     chrX 123845565-123845574      *
#>   -------
#>   seqinfo: 40 sequences (1 circular) from canFam3 genome
#> 
#> $danRer7.masked
#> GRanges object with 24170 ranges and 0 metadata columns:
#>           seqnames            ranges strand
#>              <Rle>         <IRanges>  <Rle>
#>       [1]     chr1     508054-508153      *
#>       [2]     chr1     509073-509172      *
#>       [3]     chr1     509982-510081      *
#>       [4]     chr1     512882-512981      *
#>       [5]     chr1     513539-513638      *
#>       ...      ...               ...    ...
#>   [24166]    chr25 38489453-38489552      *
#>   [24167]    chr25 38492045-38492144      *
#>   [24168]    chr25 38493834-38493933      *
#>   [24169]    chr25 38495078-38495177      *
#>   [24170]    chr25 38498098-38498197      *
#>   -------
#>   seqinfo: 26 sequences (1 circular) from danRer7 genome
#> 
#> $galGal4.masked
#> GRanges object with 9258 ranges and 0 metadata columns:
#>                       seqnames        ranges strand
#>                          <Rle>     <IRanges>  <Rle>
#>      [1]                  chr1     1034-1133      *
#>      [2]                  chr1     1635-1734      *
#>      [3]                  chr1     2505-2604      *
#>      [4]                  chr1     3124-3223      *
#>      [5]                  chr1     4729-4828      *
#>      ...                   ...           ...    ...
#>   [9254] chrLGE22C19W28_E50C23 923755-937231      *
#>   [9255] chrLGE22C19W28_E50C23 939390-939489      *
#>   [9256] chrLGE22C19W28_E50C23 940663-940900      *
#>   [9257] chrLGE22C19W28_E50C23 942958-943949      *
#>   [9258] chrLGE22C19W28_E50C23 961456-962294      *
#>   -------
#>   seqinfo: 34 sequences (1 circular) from galGal4 genome
#> 
#> $hg19.masked
#> GRanges object with 460 ranges and 0 metadata columns:
#>                   seqnames          ranges strand
#>                      <Rle>       <IRanges>  <Rle>
#>     [1]               chr1         1-10000      *
#>     [2]               chr1   177418-227417      *
#>     [3]               chr1   267720-317719      *
#>     [4]               chr1   471369-521368      *
#>     [5]               chr1 2634221-2684220      *
#>     ...                ...             ...    ...
#>   [456] chr19_gl949751_alt   789619-815618      *
#>   [457] chr19_gl949752_alt   253363-258362      *
#>   [458] chr19_gl949753_alt   195632-213631      *
#>   [459] chr19_gl949753_alt   361078-363077      *
#>   [460] chr19_gl949753_alt   509415-609414      *
#>   -------
#>   seqinfo: 298 sequences (2 circular) from hg19 genome
#> 
#> $hg38.masked
#> GRanges object with 1024 ranges and 0 metadata columns:
#>                      seqnames          ranges strand
#>                         <Rle>       <IRanges>  <Rle>
#>      [1]                 chr1         1-10000      *
#>      [2]                 chr1   207667-257666      *
#>      [3]                 chr1   297969-347968      *
#>      [4]                 chr1   535989-585988      *
#>      [5]                 chr1 2702782-2746290      *
#>      ...                  ...             ...    ...
#>   [1020] chr19_KV575258v1_alt   106475-106477      *
#>   [1021] chr19_KV575258v1_alt   106479-106481      *
#>   [1022] chr19_KV575258v1_alt   106484-106485      *
#>   [1023] chr19_KV575260v1_alt     62709-62710      *
#>   [1024] chr19_KV575260v1_alt     70481-70482      *
#>   -------
#>   seqinfo: 595 sequences (1 circular) from hg38 genome
#> 
#> $mm10.masked
#> GRanges object with 587 ranges and 0 metadata columns:
#>               seqnames            ranges strand
#>                  <Rle>         <IRanges>  <Rle>
#>     [1]           chr1         1-3000000      *
#>     [2]           chr1   3002130-3003118      *
#>     [3]           chr1 22424868-22424967      *
#>     [4]           chr1 22433269-22466510      *
#>     [5]           chr1 75055557-75121556      *
#>     ...            ...               ...    ...
#>   [583] chrUn_GL456393       53788-53887      *
#>   [584] chrUn_GL456394         1035-2612      *
#>   [585] chrUn_GL456394       20897-21977      *
#>   [586] chrUn_GL456396         7730-8220      *
#>   [587] chrUn_GL456396       11697-11940      *
#>   -------
#>   seqinfo: 66 sequences (1 circular) from mm10 genome
#> 
#> $mm9.masked
#> GRanges object with 1025 ranges and 0 metadata columns:
#>              seqnames            ranges strand
#>                 <Rle>         <IRanges>  <Rle>
#>      [1]         chr1         1-3000000      *
#>      [2]         chr1 22414949-22415048      *
#>      [3]         chr1 22423350-22473349      *
#>      [4]         chr1 24686639-24736638      *
#>      [5]         chr1 75102131-75118130      *
#>      ...          ...               ...    ...
#>   [1021] chrUn_random   5884087-5884186      *
#>   [1022] chrUn_random   5892169-5892268      *
#>   [1023] chrUn_random   5894720-5896266      *
#>   [1024] chrUn_random   5897640-5897925      *
#>   [1025] chrUn_random   5898884-5899265      *
#>   -------
#>   seqinfo: 35 sequences (1 circular) from mm9 genome
#>