Analyze sample noise

Analyzes the noiseness of the supplied sample using the method described by Diaz et al.. See details.

noiseAnalyze(object, windows = NULL, force = FALSE)

Arguments

object

An RLRanges object.

windows

Genomics windows to use for quantifying signal. Will be automatically supplied if not provided. It is recommended NOT to specify this option for most analysis types, as doing so will impair the ability to compare to RLBase samples. Default: NULL.

force

Force noiseAnalyze to run, even if on Windows. Default: FALSE.

Value

An RLRanges object with noise analysis results included as a tbl. The result is accessed via rlresults(object, "noiseAnalysis").

Details

Currently, this does not work on windows.

Method

The method used for noise analysis is a minor modification of the method developed by Diaz et al., 2012 and also implemented by the deepTools function, plotFingerprint.

Briefly, if user-supplied RLRanges contain a bigWig coverage file, then the coverage is quantified within random genomic regions (randomWindows). The regions are then ranked. A good signal-to-noise ratio will yield a distribution where most bins have little coverage but a few have very high coverage. Use downstream tools like plotNoise and plotCompareNoise to visualize these results.

Examples

# Example RLRanges object
rlr <- readRDS(system.file("extdata", "rlrsmall.rds", package = "RLSeq"))

# noiseAnalyze does not work on Windows OS
if (.Platform$OS.type != "windows") {
    # run noiseAnalyze
    rlr <- noiseAnalyze(rlr)
}