Random Genomic Windows

A collection of random genomic windows for use with noiseAnalyze. See the data-raw/genome_masks.R script for the processing steps.

randomWindows

Format

An object of class list of length 3.

Details

Structure

randomWindows is a named list of tbl objects containing ~1000 random genomic regions. The names are the genomes to which the regions correspond.

Columns (mirrors bed3 format).

chrom - the chromosome in UCSC format

start - the starting position

end - the end position

Examples

randomWindows
#> $mm10
#> # A tibble: 1,000 × 3
#>    chrom    start      end
#>    <chr>    <int>    <int>
#>  1 chr1     48578    49578
#>  2 chr1   1939982  1940982
#>  3 chr1   2712979  2713979
#>  4 chr1   3554553  3555553
#>  5 chr1   4524757  4525757
#>  6 chr1   6759610  6760610
#>  7 chr1   7502531  7503531
#>  8 chr1   8046629  8047629
#>  9 chr1   9032815  9033815
#> 10 chr1  11530044 11531044
#> # … with 990 more rows
#> # ℹ Use `print(n = ...)` to see more rows
#> 
#> $hg38
#> # A tibble: 984 × 3
#>    chrom    start      end
#>    <chr>    <int>    <int>
#>  1 chr1     60009    61009
#>  2 chr1   2396448  2397448
#>  3 chr1   3351327  3352327
#>  4 chr1   3950461  3951461
#>  5 chr1   4390919  4391919
#>  6 chr1   5589406  5590406
#>  7 chr1   8350107  8351107
#>  8 chr1   9811405  9812405
#>  9 chr1   9939954  9940954
#> 10 chr1  11158183 11159183
#> # … with 974 more rows
#> # ℹ Use `print(n = ...)` to see more rows
#> 
#> $hg19
#> # A tibble: 999 × 3
#>    chrom    start      end
#>    <chr>    <int>    <int>
#>  1 chr1     60009    61009
#>  2 chr1   2396448  2397448
#>  3 chr1   2738078  2739078
#>  4 chr1   3557198  3558198
#>  5 chr1   4390919  4391919
#>  6 chr1   5589406  5590406
#>  7 chr1   9811405  9812405
#>  8 chr1   9939954  9940954
#>  9 chr1  11158183 11159183
#> 10 chr1  13045451 13046451
#> # … with 989 more rows
#> # ℹ Use `print(n = ...)` to see more rows
#>