RLBase Sample Transcript Feature Overlaps

Summary statistics from transcript feature overlap analysis of peaks from all RLBase samples.

rlsampleTxOl

Format

An object of class tbl_df (inherits from tbl, data.frame) with 4754 rows and 4 columns.

Details

Structure

rlsampleTxOl is a tbl with the following columns:

rlsample

The RLBase sample identifier for the sample.

Matches the rlsample column in RLHub::rlbase_samples

feature

The transcript feature for which overlap analysis was performed.

These features were derived from the Transcript Features collection described in RLHub::annotations

n

The raw number of peaks from the sample overlapping a feature.

pct

The proportion of peaks fro the sample overlapping a feature.

Examples

rlsampleTxOl
#> # A tibble: 4,754 × 4
#>    rlsample   feature        n    pct
#>    <chr>      <chr>      <int>  <dbl>
#>  1 SRX1070676 Exon        3379 0.0942
#>  2 SRX1070676 fiveUTR      601 0.0168
#>  3 SRX1070676 Intergenic  3780 0.105 
#>  4 SRX1070676 Intron     21991 0.613 
#>  5 SRX1070676 threeUTR    1147 0.0320
#>  6 SRX1070676 TSS         2884 0.0804
#>  7 SRX1070676 TTS         2071 0.0578
#>  8 SRX1070677 Exon        2422 0.104 
#>  9 SRX1070677 fiveUTR      474 0.0203
#> 10 SRX1070677 Intergenic  2488 0.107 
#> # … with 4,744 more rows
#> # ℹ Use `print(n = ...)` to see more rows