Auxiliary Data

A list containing data used by RLSeq functions. It can also be useful for checking the available modes and genomes in RLSeq. See also the data-raw/auxdata.R script that was used to create it.

auxdata

Format

An object of class list of length 12.

Details

Structure

A named list containing the following entries:

db_cols

A tbl with colors associated with each database in RLHub useful for plotting. See also RLHub::annotations.

annotypes

A tbl containing the annotation databases and annotation types available from RLBase. See also RLHub::annotations.

ip_cols

A tbl containing the colors associated with each "Immunoprecipitation type" (ip_type) in RLBase. See also RLHub::rlbase_samples.

mode_cols

A tbl containing the colors associated with each R-loop mapping mode in RLHub::rlbase_samples.

heat_cols

A tbl containing the colors associated with user-supplied data and RLBase data when running corrHeatmap.

label_cols

A tbl containing the colors associated with the labels in RLBase. See also RLHub::rlbase_samples.

prediction_cols

A tbl containing the colors associated with the predictions in RLBase. See also RLHub::rlbase_samples.

prediction_label_cols

A tbl containing the colors associated with the prediction-label combinations in RLBase. See also RLHub::rlbase_samples.

available_modes

A tbl containing the modes available in RLBase and associated metadata. See also RLHub::rlbase_samples.

available_genomes

A character showing all the official UCSC genomes available for use with RLSeq. See also available_genomes.

misc_modes

A character showing the R-loop mapping modes that are lumped into the 'misc' category for simplification of plotting.

Examples

auxdata
#> $db_cols
#> # A tibble: 20 × 2
#>    db                  col    
#>    <chr>               <chr>  
#>  1 Centromeres         #876800
#>  2 CNA                 #9546C0
#>  3 Cohesin             #BA3D5D
#>  4 CpG_Islands         #00878E
#>  5 Encode_CREs         #AD34B0
#>  6 Encode_Histone      #BD317E
#>  7 encodeTFBS          #B44741
#>  8 G4Qexp              #9A5E00
#>  9 G4Qpred             #007AB8
#> 10 knownGene_RNAs      #006CC4
#> 11 Microsatellite      #6F7100
#> 12 PolyA               #0082A6
#> 13 RBP_ChIP            #A95300
#> 14 RBP_eCLiP           #6A5AC7
#> 15 Repeat_Masker       #007E00
#> 16 skewr               #4D7800
#> 17 snoRNA_miRNA_scaRNA #008650
#> 18 Splice_Events       #BA2C9A
#> 19 Transcript_Features #00831F
#> 20 tRNAs               #008872
#> 
#> $annotypes
#> # A tibble: 638 × 2
#>    db          type       
#>    <chr>       <chr>      
#>  1 Centromeres Centromeres
#>  2 CNA         0          
#>  3 CNA         1          
#>  4 CNA         2          
#>  5 CNA         3          
#>  6 CNA         4          
#>  7 Cohesin     STAG1      
#>  8 Cohesin     STAG2      
#>  9 CpG_Islands CpG_Islands
#> 10 Encode_CREs CTCF       
#> # … with 628 more rows
#> # ℹ Use `print(n = ...)` to see more rows
#> 
#> $rlbps
#> # A tibble: 788 × 5
#>    geneName NADEL.2015 WANG.2018 CRISTINI.2018 combinedScore
#>    <chr>         <dbl>     <dbl>         <dbl>         <dbl>
#>  1 DDX21          0.9     0              0.968          1.87
#>  2 DHX9           0.85    0.0635         0.831          1.74
#>  3 HNRNPR         1       0              0.731          1.73
#>  4 ILF3           0.85    0.0708         0.749          1.67
#>  5 NKRF           0       1              0.636          1.64
#>  6 PDCD11         0       1              0.633          1.63
#>  7 PWP1           0       1              0.620          1.62
#>  8 SRPK1          0       1              0.457          1.46
#>  9 PWP2           0       0.960          0.472          1.43
#> 10 NPM3           0       1              0.397          1.40
#> # … with 778 more rows
#> # ℹ Use `print(n = ...)` to see more rows
#> 
#> $ip_cols
#> # A tibble: 3 × 3
#>   ip_type POS     NEG    
#>   <chr>   <chr>   <chr>  
#> 1 S9.6    #2E294E #2E2C3B
#> 2 dRNH    #C1292E #955253
#> 3 Other   #F3B700 #D8BC94
#> 
#> $mode_cols
#> # A tibble: 11 × 2
#>    mode      col    
#>    <chr>     <chr>  
#>  1 MapR      #FE7267
#>  2 qDRIP     #DC8E00
#>  3 ssDRIP    #A3A600
#>  4 RDIP      #24B700
#>  5 sDRIP     #00C27A
#>  6 DRIPc     #00C2C8
#>  7 DRIP      #00B2FF
#>  8 RLCnT-S96 #938DFF
#>  9 RLCnT-HBD #EF62FB
#> 10 R-ChIP    #FF58BF
#> 11 misc      #8f8f8f
#> 
#> $heat_cols
#> # A tibble: 2 × 2
#>   selected      col    
#>   <chr>         <chr>  
#> 1 RLBase        #e0dede
#> 2 user_selected #2c408a
#> 
#> $label_cols
#> # A tibble: 2 × 2
#>   label col    
#>   <chr> <chr>  
#> 1 POS   #e0dede
#> 2 NEG   #8a2c2c
#> 
#> $prediction_cols
#> # A tibble: 2 × 2
#>   prediction col    
#>   <chr>      <chr>  
#> 1 POS        #e0dede
#> 2 NEG        #8a2c2c
#> 
#> $prediction_label_cols
#>   NEG_NEG   NEG_POS   POS_NEG   POS_POS 
#> "#8a2c2c" "#A76767" "#7aa4c4" "#2270ab" 
#> 
#> $available_modes
#> # A tibble: 20 × 6
#>    mode         family ip_type strand_specific moeity bisulfite_seq
#>    <chr>        <chr>  <chr>   <lgl>           <chr>  <lgl>        
#>  1 DRIPc        DRIP   S9.6    TRUE            RNA    FALSE        
#>  2 DRIP         DRIP   S9.6    FALSE           DNA    FALSE        
#>  3 sDRIP        DRIP   S9.6    TRUE            DNA    FALSE        
#>  4 ssDRIP       DRIP   S9.6    TRUE            DNA    FALSE        
#>  5 DRNA         Other  None    FALSE           RNA    FALSE        
#>  6 MapR         MapR   dRNH    FALSE           DNA    FALSE        
#>  7 m6A-DIP      RDIP   RNA-m6A FALSE           DNA    FALSE        
#>  8 DRIVE        Other  dRNH    FALSE           RNA    FALSE        
#>  9 RLCnT-HBD    RLCnT  dRNH    TRUE            DNA    FALSE        
#> 10 RLCnT-S96    RLCnT  S9.6    TRUE            DNA    FALSE        
#> 11 DRIPc-HBD    DRIP   dRNH    TRUE            RNA    FALSE        
#> 12 qDRIP        DRIP   S9.6    TRUE            DNA    FALSE        
#> 13 RDIP         RDIP   S9.6    TRUE            RNA    FALSE        
#> 14 DRIP-RNA-Seq DRIP   S9.6    TRUE            RNA    FALSE        
#> 15 DREAM        Other  None    TRUE            RNA    FALSE        
#> 16 R-ChIP       R-ChIP dRNH    TRUE            DNA    FALSE        
#> 17 RR-ChIP      R-ChIP dRNH    TRUE            RNA    FALSE        
#> 18 RNH-CnR      MapR   dRNH    FALSE           DNA    FALSE        
#> 19 S1-DRIP      DRIP   S9.6    FALSE           DNA    FALSE        
#> 20 BisMapR      MapR   dRNH    TRUE            DNA    FALSE        
#> 
#> $available_genomes
#>  [1] "canFam3"  "ce10"     "ce11"     "ce6"      "ci3"      "danRer10"
#>  [7] "danRer11" "danRer7"  "dm6"      "fr3"      "galGal4"  "galGal5" 
#> [13] "galGal6"  "gorGor3"  "gorGor4"  "gorGor5"  "hg19"     "hg38"    
#> [19] "mm10"     "mm9"      "panPan1"  "panTro3"  "panTro4"  "panTro5" 
#> [25] "panTro6"  "papAnu2"  "rheMac3"  "rheMac8"  "rn5"      "rn6"     
#> [31] "sacCer2"  "sacCer3"  "susScr11" "susScr3"  "taeGut2"  "tetNig2" 
#> [37] "xenTro3"  "xenTro7"  "xenTro9" 
#> 
#> $misc_modes
#>  [1] "BisMapR"      "DREAM"        "DRIP-RNA-Seq" "DRIPc-HBD"    "DRIVE"       
#>  [6] "DRNA"         "m6A-DIP"      "RNH-CnR"      "RR-ChIP"      "S1-DRIP"     
#>