Tests Genomic Feature Enrichment

Tests the enrichment of genomic features in supplied peaks.

featureEnrich(
  object,
  annotype = c("primary", "full"),
  annotations = NULL,
  downsample = 10000,
  quiet = FALSE
)

Arguments

object
An RLRanges object.

annotype
The type of annotations to use. Options include "primary" (missing ChIP and eCLiP data) and "full" (all annotations). Default: "primary".

annotations
Annotation list. See details.

downsample
If a numeric, data will be down sampled to the requested number of peaks. This improves the speed of genomic shuffling and helps prevent p-value inflation. If FALSE, then downsampling will not be performed. Default: 10000.

quiet
If TRUE, messages will be suppressed. Default: FALSE the same format as RLSeq::annotationLst.

Value

An RLRanges object containing the results of the enrichment test.

Details

annotations

RLHub: RLHub::annots_full_hg38(), for example, will return the full suite of annotations for the hg38 genome.

Custom: For custom annotations, the only requirement is that annotations is a named list of tbls in which each tbl follows the structure:

chrom	start	end	strand	name

chr1	10015	10498	+	skewr__C_SKEW__1

chr1	10614	11380	+	skewr__G_SKEW__1

Such a list can be generated directly from RLBase S3 bucket files:

  small_anno <- list(
    "Centromeres" = readr::read_csv(
      paste0(
          "https://rlbase-data.s3.amazonaws.com",
          "/annotations/hg38/Centromeres.csv.gz"
      )
    ),
    "SkewR" = readr::read_csv(
      paste0(
          "https://rlbase-data.s3.amazonaws.com/",
          "annotations/hg38/SkewR.csv.gz")
    )
  )

Examples

# Example RLRanges dataset
rlr <- readRDS(system.file("ext-data", "rlrsmall.rds", package = "RLSeq"))

# RL Region Test
featureEnrich(rlr)

# With custom annotations
small_anno <- list(
    "Centromeres" = readr::read_csv(
        paste0(
            "https://rlbase-data.s3.amazonaws.com/",
            "annotations/hg38/Centromeres.csv.gz"
        ),
        show_col_types = FALSE
    ),
    "SkewR" = readr::read_csv(
        paste0(
            "https://rlbase-data.s3.amazonaws.com/",
            "annotations/hg38/SkewR.csv.gz"
        ),
        show_col_types = FALSE
    )
)
featureEnrich(rlr, annotations = small_anno)